backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.118.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.0
filelock==3.19.1
fonttools==4.60.0
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.9.1.4
nvidia-cudnn-cu12==9.13.1.26
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
polars==1.33.1
psutil==7.1.0
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.7.0
scipy==1.16.2
sentry-sdk==2.39.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
sympy==1.14.0
torch==2.8.0+cu126
torchvision==0.23.0+cu126
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.3.203
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
